:hammer: Move Realm storage initialization to commonMain
